Store filenames as copy-on-write `SharedPtr` with `std::string` instead of `CompactString`. This makes copying unowning handles more efficient and is simpler to implement, at the cost of more allocations in some cases.
Change `filename` parameters of opening functions from `PathRef` to
`PathInitializer` (except for `OpenAt()` and related functions where
materializing the filename as `std::string` is not useful). The parameters can
benefit from moving from `std::string` (as it once used to be), and this could
be easily done since `PathInitializer` exists now.

This change makes it easier to add further metadata besides the filename in the
future.

PiperOrigin-RevId: 781940463