Make `ABSL_NULLABILITY_COMPATIBLE` effective by adding private `pointer` member type aliases.
Constraints for `ABSL_NULLABILITY_COMPATIBLE` to be effective are guessed from
experiments:

`pointer` must be a raw pointer type (or `element_type` must be defined).
It must be defined directly in the class, not inherited. It can be private.

Passing an `absl_nullable` value or a literal `nullptr` to an `absl_nonnull`
parameter is caught. An object constructed from a nullable value is assumed to
be nullable. `operator*` requires a non-null object.

If `pointer` matches the result type of `get()`, passing the result of `get()`
to an `absl_nonnull` parameter is caught, and `operator->` requires a non-null
object. Otherwise these aspects are not checked. Hence if there is a suitable
pointer type for `get()` and `operator->`, it is used, otherwise `pointer` can
be an arbitrary pointer type.

Also, add some missing `ABSL_NULLABILITY_COMPATIBLE` annotations.

PiperOrigin-RevId: 795568636