Let `CsvReader` recognize only CR-LF or LF as the newline, but not CR: like `ReadNewline::kCrLfOrLf`, not `ReadNewline::kAny`. CR without LF is treated as an error, except that it can be quoted.
CR as a newline is not used by any modern system. It was used by classic Mac OS
(unsupported since 2002), 8-bit computers, and MIT Lisp Machines. It is doubtful
that such files are encountered today.

RFC4180 specifies only CR-LF, recognizing LF is an extension.

This is consistent with the removal of `ReadNewline::kAny`.

PiperOrigin-RevId: 796917900