If `Limiting{Reader,{,Backward}Writer}Base::Options::set_max_{pos,length}()` are both used, let the last call overwrite the previous state, instead of making that a precondition violation.
Remove the special case of `ScopedLimiter` that if `options.exact()` is `true`
then a new limit must be set. The exact size cannot possibly be reached, and
`LimitingReader` does not treat `exact()` without a limit specially, but fails
at the end. This behavior is preserved, with the usual `ScopedLimiter` change
that failure is immediate, even if this behavior is unlikely to be useful.

Let `Limiting{,Backward}Writer::MakeBuffer()` ensure the invariant that
`start_pos() <= max_pos_`. This is important during initialization and
`SetMaxPos()`, freeing the caller from having to ensure that `max_pos()`
is not already exceeded.

Minor code cleanups. Do not be so crazy about inlining.

PiperOrigin-RevId: 808516773