Change `SerializedMessageOptions::deterministic()` from `bool` to `std::optional<bool>`. `std::nullopt` means `proto2::io::CodedOutputStream::IsDefaultSerializationDeterministic()`.
This avoids calling
`proto2::io::CodedOutputStream::IsDefaultSerializationDeterministic()` multiple
times to set the default and then to check if the setting is the default.
That function reads an atomic global variable so it is not free.

For simplicity, checking whether the setting is the default compares
`deterministic()` only against `std::nullopt`. If it is explicitly set,
it is very unlikely that it matches
`proto2::io::CodedOutputStream::IsDefaultSerializationDeterministic()`
(that is usually `false`, while an explicit setting is usually `true`).
The comparison can be conservative: the default case is an optimization,
the non-default case can always be used.

PiperOrigin-RevId: 808886682