Add back `SerializedMessage{Reader,Rewriter}::OnLengthUnchecked()`.
It turns out to be significantly faster after all, the `ScopedLimiter` overhead
is too high.

This time it includes a check there against the source length, to make
processing untrusted input safer. In `OnLengthDelimited()` the analogous check
is performed by `ScopedLimiter` with the `exact` parameter.

Simplify `ScopedLimiter` constructor: use `LimitingReader::max_length()` and
`LimitingReader::set_max_length()` instead of converting the length to the
position first (which may overflow) and then working in terms of the position.
This delegates an overflow check to `set_max_length()`.

Tweak `Limiting{Reader,Writer,BackwardWriter}` failure messages to report
lengths instead of positions when appropriate.

PiperOrigin-RevId: 811779742