Fix uses of std::forward.
`std::forward<T>(variable)` should spell the declared type `T` of the variable, exactly as it was originally spelled in the variable's declaration, except with any trailing `&&` stripped.
This guarantees that the variable is always forwarded with its original value category (lvalue-ness or rvalue-ness), which is the purpose of `std::forward`.

PiperOrigin-RevId: 814559135