Let `SerializedMessageRewriter` commit unchanged fields after processing the message.
When there are no field-specific actions registered for `SerializedMessageRewriter`, it should copy the entire source message to the destination.

`SerializedMessageRewriter` delays copying unchanged fields in order to copy a region of them in one go. This change ensures that at the end of `Rewrite()`, the pending unchanged region is flushed to the destination.

PiperOrigin-RevId: 819311559