Add `WireType::kInvalid6` and `WireType::kInvalid7`.
This allows writing `switch` statements over the wire type which are more
efficient.

Wire types are normally decoded from 3 bits of a field tag. Values 0 to 5 are in
use. Values 6 and 7 are invalid, while larger values are impossible.

With a `default` case or falling to the end of a `switch`, the compiler does not
necessarily realize that only 8 cases are possible, and adds a bounds check
which covers values 6 and 7 as well as larger values.

With `RIEGELI_ASSUME_UNREACHABLE()` after the `switch`, which promises that
larger values are impossible, the compiler-generated jump table has two
additional entries, while the bounds check is skipped.

Cosmetics: reorder `WireType` enumerators so that their order corresponds
to the order in which they are typically handled. This moves `kFixed32` up;
for historical reasons, it has a high wire type number.
PiperOrigin-RevId: 820562774