Add `ReaderSpan` to specify a span of `Reader` contents from the current position with the given length.
It cooperates with `LimitingReader` to make processing more efficient using
`ScopedLimiter`, and with `Dependency` so that it can be passed to all functions
expecting an object providing a `Reader`.

For now it replaces `ParseMessageOfLength()` with a specialization of
`ParseMessage()`, and an overload of
`SerializedMessage{,Backward}Writer::CopyString()` with a specialization of
another one. It will have other uses soon.

PiperOrigin-RevId: 825871554