Fix an embarrassing thread safety bug in `Reader::NewReader()`.
`NewReader()` promises that it can be called concurrently with other operations
(as long as `SupportsNewReader()`). This is relied upon elsewhere.

This was unfortunately not true after some classes employed an optimization of
sharing the current buffer with the new `Reader` (Jan 2022).

Luckily `NewReader()` has an overload which starts from the current position
rather than an explicitly specified position (since Aug 2022). This overload
cannot be thread-safe anyway, because the current position may change at any
time.

This overload is being renamed to `NewReaderCurrentPos()` because it differs not
only by the position being implicit:

* If `SupportsNewReader()` is `true`, then `NewReader()` is effectively const
  and thread-safe. It may be called concurrently with other operations, even
  non-const ones.

* If `SupportsNewReader()` is `true`, then `NewReaderCurrentPos()` is
  effectively const but not thread-safe. It may be called concurrently
  with const operations and with other calls to `NewReader()` and
  `NewReaderCurrentPos()`, but not with non-const operations.

Clean up comments at `ReaderFactory::NewReader()`:

In contrast to `Reader::NewReader()` and `Reader::NewReaderCurrentPos()`,
`ReaderFactory::NewReader()` is unconditionally const and thread-safe,
even with an implicit initial position (concurrency with other operations
is not applicable because the original `Reader` must not be accessed
directly). The optimization of sharing the current buffer is applicable
as long as the initial position is implicit or matches `pos()`.

PiperOrigin-RevId: 828515012