Do not require to call `SkipLengthDelimited()`. Let `riegeli::field_handlers::OnLengthDelimited()` handle that automatically.
Remembering about that was subtle, and it would have become even more subtle
given some future changes.

In the field handler protocol, `HandleLengthDelimitedFromReader()` is still
responsible for skipping to the end. This is ensured only on the level of
`riegeli::field_handlers::OnLengthDelimited()`.

`SkipLengthDelimitedFromReader()` is retained for implementing the field handler
protocol.

The cost is performing redundant operations for most field handlers which do not
need to force skipping to the end. Oh well.

PiperOrigin-RevId: 839610885