In the field handler protocol, distinguish `Handle` functions for dynamic field numbers by naming them with the `DynamicHandle` prefix.
This makes the interpretation of the signature of the given handling funtion
unambiguous. While in practice the functions are called only when the role has
been determined by `kFieldNumber` and by the presence of `Accept` functions,
it is cleaner when the signature can be interpreted on its own: which parameter
is `accepted` if it exists, which parameter is `repr`, and which parameters are
`context...`.

PiperOrigin-RevId: 842793312