Include `<stddef.h>` instead of `<cstddef>` (except for `std::nullptr_t`) and `<stdint.h>` instead of `<cstdint>`.
This was the original convention but some cases slipped through.

These headers guarantee type names being present in the global namespace, not
only in `std::`. While in practice they are interchangeable, I see no reason to
be non-portable in this respect, even though the majority of Google code does
not care.

PiperOrigin-RevId: 847713149