In `SerializedMessage{,Backward}Writer`, rename `CopyString()` to `WriteString()`, merging it with other overloads.
Add direct support for `CordIteratorSpan` in `WriteString()`. It is convertible
to `absl::Cord`, but the stringification mechanism does not find that.

Rationale: it is not that important to explicitly indicate whether reading is
destructive or not, especially since `ReaderSpan` is move-only. It is more
important to be able to treat `ReaderSpan`, `CordIteratorSpan`, and
`absl::string_view` uniformly.
PiperOrigin-RevId: 847835642