Let `SerializedMessageReader2` use the `riegeli::TargetT` mechanism for initializing field handlers instead of `std::decay_t`. This allows to use `riegeli::Maker()` for constructing them in-place.
Make `SerializedMessageReader2` a value type in Riegeli style, with default
constructor and `Reset()`, to enable late initialization.

Raw pointers to proper field handlers are no longer automatically treated
as field handlers. References are supported instead, which is more natural,
because calling member functions on references is not special.

With `riegeli::TargetT`, `std::cref(field_handler)` deduces a reference type as
the type to store. This replaces a raw pointer.

In a future change, this will make using `FieldHandlerMap` easier, often
constructing it in-place instead of using `std::cref()`.

PiperOrigin-RevId: 847862733